+++ pfs-accounting/src/main/java/ru/gb/dev/spring/pfs/accounting/model/entity/base/AbstractNamedEntity.java
@@ -16,8 +16,8 @@ import javax.persistence.MappedSuperclass;
 @AllArgsConstructor
 public abstract class AbstractNamedEntity extends AbstractBaseEntity {
 
-	@NotNull
-	@Column(name = "name")
-	private String name = "";
+    @NotNull
+    @Column(name = "name")
+    private String name = "";
 
 }
