+++ pfs-accounting/src/main/java/ru/gb/dev/spring/pfs/accounting/security/JwtTokenStoreConfig.java
@@ -1,4 +1,4 @@
-package ru.gb.dev.spring.pfs.accounting.system.security;
+package ru.gb.dev.spring.pfs.accounting.security;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -7,7 +7,7 @@ import org.springframework.security.oauth2.provider.token.DefaultTokenServices;
 import org.springframework.security.oauth2.provider.token.TokenStore;
 import org.springframework.security.oauth2.provider.token.store.JwtAccessTokenConverter;
 import org.springframework.security.oauth2.provider.token.store.JwtTokenStore;
-import ru.gb.dev.spring.pfs.accounting.system.ServiceConfig;
+import ru.gb.dev.spring.pfs.accounting.config.ServiceConfig;
 
 /**
  * @author V. Zinchenko
