+++ pfs-accounting/src/main/java/ru/gb/dev/spring/pfs/accounting/util/UserContextFilter.java
@@ -1,4 +1,4 @@
-package ru.gb.dev.spring.pfs.accounting.util.user_context;
+package ru.gb.dev.spring.pfs.accounting.util;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
