+++ pfs-ui/src/main/java/ru/gd/dev/spring/pfs/ui/controller/dto/FailDto.java
@@ -2,8 +2,8 @@ package ru.gd.dev.spring.pfs.ui.controller.dto;
 
 public class FailDto extends ResultDto {
 
-	public FailDto(final Exception e) {
-		setSuccess(false);
-		setMessage(e.getMessage());
-	}
+    public FailDto(final Exception e) {
+        setSuccess(false);
+        setMessage(e.getMessage());
+    }
 }
