+++ pfs-ui/src/main/java/ru/gd/dev/spring/pfs/ui/controller/dto/SuccessDto.java
@@ -2,8 +2,8 @@ package ru.gd.dev.spring.pfs.ui.controller.dto;
 
 public class SuccessDto extends ResultDto {
 
-	{
-		setSuccess(true);
-	}
+    {
+        setSuccess(true);
+    }
 
 }
