+++ pfs-ui/src/main/java/ru/gd/dev/spring/pfs/ui/dto/OperationDto.java
@@ -1,0 +1,56 @@
+package ru.gd.dev.spring.pfs.ui.dto;
+
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+import lombok.NoArgsConstructor;
+import org.jetbrains.annotations.NotNull;
+import ru.gd.dev.spring.pfs.ui.dto.base.AbstractBaseDto;
+
+import javax.xml.bind.annotation.XmlRootElement;
+import java.time.LocalDateTime;
+
+@Data
+@EqualsAndHashCode(callSuper = true)
+@NoArgsConstructor
+@XmlRootElement
+public class OperationDto extends AbstractBaseDto {
+
+    private static final long serialVersionUID = -6268060818229181878L;
+
+    @NotNull
+    private String amount = "";
+
+    @NotNull
+    private String description = "";
+
+    @NotNull
+    private String comment = "";
+
+    @NotNull
+    private String type = "";
+
+    @NotNull
+    private String number = "";
+
+    @NotNull
+    private LocalDateTime operationDate = LocalDateTime.now();
+
+    @NotNull
+    private Boolean active = Boolean.FALSE;
+
+    @NotNull
+    private String clientId = "";
+
+    @NotNull
+    private String categoryId = "";
+
+    @NotNull
+    private String logoId = "";
+
+    @NotNull
+    private String accountId = "";
+
+    @NotNull
+    private String userId = "";
+
+}
