--- pfs-ui/src/main/java/ru/gd/dev/spring/pfs/ui/model/service/AccountService.java
@@ -1,52 +1,0 @@
-package ru.gd.dev.spring.pfs.ui.model.service;
-
-import org.springframework.cloud.openfeign.FeignClient;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import ru.gd.dev.spring.pfs.ui.controller.dto.ResultDto;
-import ru.gd.dev.spring.pfs.ui.model.dto.AccountDto;
-
-import java.util.List;
-
-import static org.springframework.http.MediaType.APPLICATION_JSON_UTF8_VALUE;
-
-@RequestMapping("/api/accounts")
-@FeignClient(name = "accounting")
-public interface AccountService {
-
-	@GetMapping(value = "/ping", produces = APPLICATION_JSON_UTF8_VALUE)
-	ResultDto ping();
-
-	@GetMapping(value = "{id}", produces = APPLICATION_JSON_UTF8_VALUE)
-	AccountDto get(@PathVariable("id") final String id);
-
-	@GetMapping(produces = APPLICATION_JSON_UTF8_VALUE)
-	List<AccountDto> getAll();
-
-	@PostMapping(
-			consumes = APPLICATION_JSON_UTF8_VALUE,
-			produces = APPLICATION_JSON_UTF8_VALUE
-	)
-	ResultDto post(final AccountDto accountDto);
-
-	@PutMapping(
-			consumes = APPLICATION_JSON_UTF8_VALUE,
-			produces = APPLICATION_JSON_UTF8_VALUE
-	)
-	ResultDto put(final AccountDto accountDto);
-
-	@DeleteMapping(
-			value = "{id}",
-			consumes = APPLICATION_JSON_UTF8_VALUE,
-			produces = APPLICATION_JSON_UTF8_VALUE
-	)
-	ResultDto delete(@PathVariable("id") final String accountId);
-
-	@DeleteMapping(produces = APPLICATION_JSON_UTF8_VALUE)
-	ResultDto deleteAll();
-
-}
