+++ pfs-ui/src/main/java/ru/gd/dev/spring/pfs/ui/view/content/fragment/AccountBox.java
@@ -7,7 +7,7 @@ import com.vaadin.flow.component.html.Image;
 import com.vaadin.flow.component.html.Label;
 import com.vaadin.flow.component.orderedlayout.HorizontalLayout;
 import org.jetbrains.annotations.Nullable;
-import ru.gd.dev.spring.pfs.ui.model.dto.AccountDto;
+import ru.gd.dev.spring.pfs.ui.dto.AccountDto;
 import ru.gd.dev.spring.pfs.ui.view.content.AccountView;
 
 /**
